pyproject.toml
[tool.poetry]
name = "project"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.3"
pillow = "^10.2.0"
djangorestframework = "^3.15.1"
django-rest-swagger = "^2.2.0"
drf-yasg = "^1.21.7"


[tool.poetry.group.dev.dependencies]
factory-boy = "^3.3.0"
pytest = "^8.1.1"
pytest-django = "^4.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
